tensor dmr

+++ src/cuda/gemm_tensorcores/tst/tensor_dmr/simpleTensorCoreGEMM.cu
@@ -296,7 +296,7 @@ __global__ void matrix_mult(half *A, half *B, int M, int N, int K, float *C) {
 
    
 
-      C[row * M + col] = __half2float(acc_real_t);
+      C[row * M + col] = (float)(acc_real_t);
       
    }
